+++ packages/compiler/src/module-resolver/module-resolver.ts
@@ -324,7 +324,7 @@ export async function resolveModule(
       type: "module",
       mainFile: resolved,
       manifest: pkg,
-      path: pkgDir,
+      path: await realpath(pkgDir),
     };
   }