+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/parameters/spread/SpreadClientBuilder.java
@@ -14,13 +14,17 @@ import io.clientcore.core.http.pipeline.HttpRedirectOptions;
 import io.clientcore.core.http.pipeline.HttpRedirectPolicy;
 import io.clientcore.core.http.pipeline.HttpRetryOptions;
 import io.clientcore.core.http.pipeline.HttpRetryPolicy;
+import io.clientcore.core.instrumentation.Instrumentation;
+import io.clientcore.core.instrumentation.SdkInstrumentationOptions;
 import io.clientcore.core.traits.ConfigurationTrait;
 import io.clientcore.core.traits.EndpointTrait;
 import io.clientcore.core.traits.HttpTrait;
 import io.clientcore.core.traits.ProxyTrait;
+import io.clientcore.core.utils.CoreUtils;
 import io.clientcore.core.utils.configuration.Configuration;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 import parameters.spread.implementation.SpreadClientImpl;
 
@@ -35,6 +39,9 @@ public final class SpreadClientBuilder implements HttpTrait<SpreadClientBuilder>
 
     @Metadata(properties = { MetadataProperties.GENERATED })
     private static final String SDK_VERSION = "version";
+
+    @Metadata(properties = { MetadataProperties.GENERATED })
+    private static final Map<String, String> PROPERTIES = CoreUtils.getProperties("parameters-spread.properties");
 
     @Metadata(properties = { MetadataProperties.GENERATED })
     private final List<HttpPipelinePolicy> pipelinePolicies;
@@ -179,7 +186,16 @@ public final class SpreadClientBuilder implements HttpTrait<SpreadClientBuilder>
     private SpreadClientImpl buildInnerClient() {
         this.validateClient();
         String localEndpoint = (endpoint != null) ? endpoint : "http://localhost:3000";
-        SpreadClientImpl client = new SpreadClientImpl(createHttpPipeline(), localEndpoint);
+        HttpInstrumentationOptions localHttpInstrumentationOptions = this.httpInstrumentationOptions == null
+            ? new HttpInstrumentationOptions()
+            : this.httpInstrumentationOptions;
+        SdkInstrumentationOptions sdkInstrumentationOptions
+            = new SdkInstrumentationOptions(PROPERTIES.getOrDefault(SDK_NAME, "UnknownName"))
+                .setSdkVersion(PROPERTIES.get(SDK_VERSION))
+                .setEndpoint(localEndpoint);
+        Instrumentation instrumentation
+            = Instrumentation.create(localHttpInstrumentationOptions, sdkInstrumentationOptions);
+        SpreadClientImpl client = new SpreadClientImpl(createHttpPipeline(), instrumentation, localEndpoint);
         return client;
     }
 
@@ -213,7 +229,8 @@ public final class SpreadClientBuilder implements HttpTrait<SpreadClientBuilder>
      */
     @Metadata(properties = { MetadataProperties.GENERATED })
     public ModelClient buildModelClient() {
-        return new ModelClient(buildInnerClient().getModels());
+        SpreadClientImpl innerClient = buildInnerClient();
+        return new ModelClient(innerClient.getModels(), innerClient.getInstrumentation());
     }
 
     /**
@@ -223,6 +240,7 @@ public final class SpreadClientBuilder implements HttpTrait<SpreadClientBuilder>
      */
     @Metadata(properties = { MetadataProperties.GENERATED })
     public AliasClient buildAliasClient() {
-        return new AliasClient(buildInnerClient().getAlias());
+        SpreadClientImpl innerClient = buildInnerClient();
+        return new AliasClient(innerClient.getAlias(), innerClient.getInstrumentation());
     }
 }