+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/payload/xml/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Package containing the classes for Xml.
+ * Sends and receives bodies in XML format.
+ */
+package payload.xml;