+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/versioning/returntypechangedfrom/implementation/ReturnTypeChangedFromClientImpl.java
@@ -13,6 +13,7 @@ import io.clientcore.core.http.models.HttpResponseException;
 import io.clientcore.core.http.models.RequestContext;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.http.pipeline.HttpPipeline;
+import io.clientcore.core.instrumentation.Instrumentation;
 import java.lang.reflect.InvocationTargetException;
 import versioning.returntypechangedfrom.ReturnTypeChangedFromServiceVersion;
 
@@ -68,15 +69,31 @@ public final class ReturnTypeChangedFromClientImpl {
     }
 
     /**
+     * The instance of instrumentation to report telemetry.
+     */
+    private final Instrumentation instrumentation;
+
+    /**
+     * Gets The instance of instrumentation to report telemetry.
+     * 
+     * @return the instrumentation value.
+     */
+    public Instrumentation getInstrumentation() {
+        return this.instrumentation;
+    }
+
+    /**
      * Initializes an instance of ReturnTypeChangedFromClient client.
      * 
      * @param httpPipeline The HTTP pipeline to send requests through.
+     * @param instrumentation The instance of instrumentation to report telemetry.
      * @param endpoint Need to be set as 'http://localhost:3000' in client.
      * @param serviceVersion Service version.
      */
-    public ReturnTypeChangedFromClientImpl(HttpPipeline httpPipeline, String endpoint,
+    public ReturnTypeChangedFromClientImpl(HttpPipeline httpPipeline, Instrumentation instrumentation, String endpoint,
         ReturnTypeChangedFromServiceVersion serviceVersion) {
         this.httpPipeline = httpPipeline;
+        this.instrumentation = instrumentation;
         this.endpoint = endpoint;
         this.serviceVersion = serviceVersion;
         this.service = ReturnTypeChangedFromClientService.getNewInstance(this.httpPipeline);
@@ -122,9 +139,12 @@ public final class ReturnTypeChangedFromClientImpl {
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
     public Response<String> testWithResponse(String body, RequestContext requestContext) {
-        final String contentType = "application/json";
-        final String accept = "application/json";
-        return service.test(this.getEndpoint(), this.getServiceVersion().getVersion(), contentType, accept, body,
-            requestContext);
+        return this.instrumentation.instrumentWithResponse("Versioning.ReturnTypeChangedFrom.test", requestContext,
+            updatedContext -> {
+                final String contentType = "application/json";
+                final String accept = "application/json";
+                return service.test(this.getEndpoint(), this.getServiceVersion().getVersion(), contentType, accept,
+                    body, updatedContext);
+            });
     }
 }