+++ packages/typespec-vscode/src/vscode-cmd/emit-code/emit-code.ts
@@ -1,3 +1,4 @@
+import { NodeSystemHost, resolveCompilerOptions } from "@typespec/compiler/internals";
 import { createHash } from "crypto";
 import { readFile, writeFile } from "fs/promises";
 import path from "path";
@@ -375,7 +376,6 @@ async function doEmit(
   }[] = [];
   try {
     for (const emitter of emitters) {
-      let outputDir = defaultEmitOutputDirInConfig;
       /*update emitter in config.yaml. */
       const emitNode = configYaml.get("emit");
       if (emitNode) {
@@ -405,9 +405,25 @@ async function doEmit(
         );
 
         await generateAnnotatedYamlFile(configYaml, emitter.package, baseDir);
-      } else {
-        outputDir = emitOutputDir as string;
       }
+    }
+    let newYamlContent = configYaml.toString();
+    newYamlContent = newYamlContent
+      .replaceAll(TO_BE_REPLACED_WITH_HASH, "# ")
+      .replaceAll(TO_BE_REPLACED_WITH_EMPTY_VALUE, "");
+    await writeFile(tspConfigFile, newYamlContent);
+
+    const [tspConfigOptions, diagnostics] = await resolveCompilerOptions(NodeSystemHost, {
+      entrypoint: mainTspFile,
+      cwd: baseDir,
+    });
+    if (diagnostics.length > 0) {
+      logger.debug(
+        "TypeSpec config diagnostics:",
+        diagnostics.map((d) => d.message),
+      );
+    }
+    for (const emitter of emitters) {
       let codeInfoStr: string = "code";
       if (emitter.kind !== EmitterKind.Unknown) {
         codeInfoStr = `${emitter.kind} code`;
@@ -415,17 +431,17 @@ async function doEmit(
       if (emitter.language) {
         codeInfoStr += ` for ${emitter.language}`;
       }
-      outputDir = outputDir
-        .replace("{project-root}", baseDir)
-        .replace("{output-dir}", `${baseDir}/tsp-output`)
-        .replace("{emitter-name}", emitter.package);
-      generations.push({ emitter: emitter, outputDir: outputDir, codeInfo: codeInfoStr });
+
+      let outputDir: string | undefined;
+      if (tspConfigOptions.options) {
+        outputDir = tspConfigOptions.options[emitter.package]["emitter-output-dir"];
+      }
+      generations.push({
+        emitter: emitter,
+        outputDir: outputDir ?? `${baseDir}/tsp-output/${emitter.package}`,
+        codeInfo: codeInfoStr,
+      });
     }
-    let newYamlContent = configYaml.toString();
-    newYamlContent = newYamlContent
-      .replaceAll(TO_BE_REPLACED_WITH_HASH, "# ")
-      .replaceAll(TO_BE_REPLACED_WITH_EMPTY_VALUE, "");
-    await writeFile(tspConfigFile, newYamlContent);
   } catch (error: any) {
     logger.error(error);
   }
@@ -433,7 +449,9 @@ async function doEmit(
   const allCodesToGenerate = generations
     .map((g) => `${g.codeInfo} under directory ${g.outputDir}`)
     .join(", ");
+
   logger.info(`Start to emit ${allCodesToGenerate}...`);
+
   const codeInfoStr = generations.map((g) => g.codeInfo).join(", ");
   return await vscode.window.withProgress<ResultCode>(
     {