+++ packages/typespec-vs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log - typespec-vs
 
+## 1.2.0
+
+No changes, version bump only.
+
 ## 1.1.0
 
 No changes, version bump only.